feat(ui/pipeline-designer): Move debug view to context bar
